# openapi.yaml
openapi: 3.0.3
info:
  title: Central Flow Collector API
  version: 4.0.0
  description: Single-organization central flow collection and flow analytics API.
servers:
  - url: /
paths:
...
  /api/v1/notifications/simulate:
    post:
      summary: Simulate a rule using the production evaluator and lifecycle
      description: Requires system.manage permission. Accepts rule, from and to; the simulation period is at most 24 hours. Seasonal rules read four historical weekly comparison windows for each hourly evaluation. No messages are sent and live alert state is unchanged. Results include withheld_evaluations, lifecycle counts and latest production-rendered HTML and Telegram preview when a delivery would occur.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rule, from, to]
              properties:
                rule: {type: object}
                from: {type: string, format: date-time}
                to: {type: string, format: date-time}
      responses:
        "200": {description: Simulation counts and message preview}
        "403": {description: Administrator permission required}
        "422": {description: Invalid simulation or unavailable historical data}
...
  /api/v1/engineering/{resource}:
    get:
      summary: Routing, QoS, NAT or flow field coverage analytics
      parameters:
        - name: resource
          in: path
          required: true
          schema: {type: string, enum: [routing, qos, nat, coverage]}
        - name: from
          in: query
          schema: {type: string, format: date-time}
        - name: to
          in: query
          schema: {type: string, format: date-time}
      responses:
        "200": {description: Bounded engineering summary}
        "422": {description: Storage aggregation unavailable}
  /api/v1/engineering/routes:
    post:
      summary: Replace the administrator-managed static prefix routing table
      requestBody:
        required: true
        content: {application/json: {schema: {type: array, items: {type: object}}}}
      responses:
        "200": {description: Routing table version}
  /api/v1/capacity/simulate:
    post:
      summary: Simulate retention and sampling changes without applying configuration
      requestBody:
        required: true
        content: {application/json: {schema: {type: object}}}
      responses:
        "200": {description: Explainable scenario estimate}
...